synced the local packages to the version required by the project.
I wanted to be able to reconfigure the harmony setting, so I tried to do Tools->harmoney configurator 3, and was told that there is mismatch between local package version and what's required by the project. So I was prompted to use Content Manager to load the manifest file in the project, and update the local package to match what's required in the project.
Lots of file has changed, but so far I haven't done any manual code change yet.