Targeting PIC32MZ2048EFG100 and compiles fine. To be test in Yorker Pro board.
All right, after the project setting change to target to EFG100 in the previous commit, when I go back to MHC, now System is showing the correct device now: EFG100.
Now it looks like the BSP package is gone from MHC. Looks promising. It senses that I am using EFG100 chip, which is not supported by the Starter Kit, so it took out the BSP.
Now Generate code, and I can see that some of the code that I manually commented out when doing it in the office is not there anymore either.

It is building correctly as well without any manual source code change.