Added some comments
- Used ICD3 debugging and step through this test code;
- For some reason, LATBbits.LATB12 = 1 does not work, but LATB=0x2000 works.
- SHA255: hash function that generates 32-bit hash output;
- AES-GCM: more secure than AES-CBC which is what I tried in macos;
- RSA: not tried yet.

But let the how program run, and the the blue LED shuts off, which means there is no error found during the entire suite of test.

No RTOS used in this project. This project looks very similar to what I have done in the SentryPro. Basically the App_task() is a state machine, similar to my SentryPro FSM. Easy to understand.

I start to like Harmony now. It;s easy to understand.

To see if a project was originally configured with harmony, you can check to see if there is a file called harmony-manifest-success.yml file.  Another tell-tale sign is the existence of the config directory under Source Files.

The workflow in working with harmony:
- MHC menu item will not shown by default;
- To get it shown, do Tools-> Embedded -> Harmony 3 Configurator
- choose your atdf file corresponding to the device you are targeting;
- now MHC menu item will show up;
- MHC -> Tools -> Clock configuration; or Pin Configuration etc...
- MHC -> generate code, which will update the file in the src/config directory;
- normal mplabx build.
- download and debug.